+++ services/preview/automation/mgmt/2021-06-22-preview/automation/models.go
@@ -6957,6 +6957,8 @@ type PrivateEndpointConnectionListResult struct {
 type PrivateEndpointConnectionProperties struct {
 	// PrivateEndpoint - Private endpoint which the connection belongs to.
 	PrivateEndpoint *PrivateEndpointProperty `json:"privateEndpoint,omitempty"`
+	// GroupIds - Gets the groupIds.
+	GroupIds *[]string `json:"groupIds,omitempty"`
 	// PrivateLinkServiceConnectionState - Connection State of the Private Endpoint Connection.
 	PrivateLinkServiceConnectionState *PrivateLinkServiceConnectionStateProperty `json:"privateLinkServiceConnectionState,omitempty"`
 }