+++ services/preview/automation/mgmt/2018-06-30-preview/automation/CHANGELOG.md
@@ -1,2 +1,18 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Funcs
+
+1. *PrivateEndpointConnection.UnmarshalJSON([]byte) error
+1. PrivateEndpointConnection.MarshalJSON() ([]byte, error)
+1. PrivateLinkServiceConnectionStateProperty.MarshalJSON() ([]byte, error)
+
+### Struct Changes
+
+#### New Structs
+
+1. PrivateEndpointConnection
+1. PrivateEndpointConnectionProperties
+1. PrivateEndpointProperty
+1. PrivateLinkServiceConnectionStateProperty